+++ store/auth/userThunks.ts
@@ -1,0 +1,81 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { apiClient } from '@/src/api/axios';
+import { AsyncThunkConfig } from '../store';
+import { supabase } from '@/utils/supabase';
+
+export const fetchUserInfo = createAsyncThunk<ApiResponse<User>, void, AsyncThunkConfig>(
+    'auth/fetchUserInfo',
+    async (_, { rejectWithValue }) => {
+        try {
+            const response = await apiClient.post<ApiResponse<User>>('user/check');
+            return response.data;
+        } catch (error: any) {
+            return rejectWithValue(error.response?.data?.message || 'Error connecting to API');
+        }
+    }
+);
+export const updateUserInfo = createAsyncThunk<null, { name: string; address: string }, AsyncThunkConfig>(
+    'auth/updateUserInfo',
+    async (data, { rejectWithValue }) => {
+        try {
+            const response = await apiClient.put<ApiResponse<User>>('user/user', {
+                name: data.name,
+                address: data.address,
+            });
+            return null;
+        } catch (error: any) {
+            return rejectWithValue(error.response?.data?.message || 'Error connecting to API');
+        }
+    }
+);
+
+export const updateUserImage = createAsyncThunk<null, void, AsyncThunkConfig>(
+    'auth/updateUserImage',
+    async (_, { getState, rejectWithValue }) => {
+        const image_url = getState().auth.user?.image_url;
+        try {
+            const response = await apiClient.put<ApiResponse<User>>('user/image', {
+                image_url,
+            });
+            return null;
+        } catch (error: any) {
+            return rejectWithValue(error.response?.data?.message || 'Error connecting to API');
+        }
+    }
+);
+export const updateUserPassword = () => {};
+
+export const uploadImage = createAsyncThunk<string, FormData, AsyncThunkConfig>(
+    'auth/uploadImage',
+    async (formData, { getState, rejectWithValue }) => {
+        try {
+            const file = formData.get('file') as File;
+            const userId = getState().auth.user?.id;
+            if (!file) throw new Error('No file provided');
+            const filePath = `${userId}/profile.jpg`;
+            const token = getState().auth.token;
+
+            const existingFilePath = getState().auth.user?.image_url;
+            if (existingFilePath) {
+                await supabase.storage.from('user').remove([filePath]);
+            }
+
+            const { error } = await supabase.storage.from('user').upload(filePath, file, {
+                cacheControl: '3600',
+                upsert: false,
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+            });
+
+            if (error) {
+                console.error('Upload failed:', error.message);
+                throw error;
+            }
+
+            return supabase.storage.from('user').getPublicUrl(filePath).data.publicUrl;
+        } catch (error: any) {
+            return rejectWithValue(error.message);
+        }
+    }
+);
